[warnings] ImageSourceFile Radiance/Wic .cpp: fix narrowing conversion warnings

+++ src/cinder/ImageSourceFileRadiance.cpp
@@ -136,7 +136,7 @@ namespace {
 float convertComponent( int expo, int val )
 {
 	float v = val / 256.0f;
-	float d = powf( 2.0f, expo );
+	float d = powf( 2.0f, static_cast<float>( expo ) );
 	return v * d;
 }
